pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/filpe"]

[project]
name = "filpe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.1",
    "injector>=0.24.0",
    "openpyxl>=3.1.5",
    "orjson>=3.11.7",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-multipart>=0.0.22",
    "rq>=2.7.0",
    "structlog>=25.5.0",
    "uvicorn>=0.42.0",
    "valkey>=6.1.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[project.scripts]
filpe = "filpe.cli:main"
